fix(og): fixed open graph field names
`:` Isn't allowed character for a field when the grapqhl plugin tries to generate definitions for it. 

This makes the build of the app fail if one has both Graphql Plugin and Seo plugin and adds an seo component to some content-type. 

Changing the names of the open-graphql components to not include the `:` fixes this bug.